requirements.txt
streamlit
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit-tags
pdfminer.six
PyPDF2
sqlite3
python-dotenv
groq